빌드 warning 수정 (#1)
* fix : warning 제거

* fix : LuaScript 파일 선택 박스

* feat : Lua 키매핑을 함수로 처리(일부 키만 지원)

* fix : 빌드 warning 수정

* fix : LuaScript관련 기존 코드 제거

+++ EngineSIU/EngineSIU/Engine/Source/Editor/LevelEditor/SLevelEditor.cpp
@@ -21,16 +21,16 @@ SLevelEditor::SLevelEditor()
 
 void SLevelEditor::Initialize(uint32 InEditorWidth, uint32 InEditorHeight)
 {
-    EditorWidth = InEditorWidth * 0.8f;
-    EditorHeight = InEditorHeight - 104.f;
+    EditorWidth = static_cast<uint32>(static_cast<float>(InEditorWidth) * 0.8f);
+    EditorHeight = InEditorHeight - 104;
     
     ResizeEditor(EditorWidth, EditorHeight);
     
     VSplitter = new SSplitterV();
-    VSplitter->Initialize(FRect(0.0f, 0.f, EditorWidth, EditorHeight));
-    
+    VSplitter->Initialize(FRect(0.0f, 0.f, static_cast<float>(EditorWidth), static_cast<float>(EditorHeight)));
+
     HSplitter = new SSplitterH();
-    HSplitter->Initialize(FRect(0.f, 0.0f, EditorWidth, EditorHeight));
+    HSplitter->Initialize(FRect(0.f, 0.0f, static_cast<float>(EditorWidth), static_cast<float>(EditorHeight)));
     
     FRect Top = VSplitter->SideLT->GetRect();
     FRect Bottom = VSplitter->SideRB->GetRect();
@@ -178,7 +178,7 @@ void SLevelEditor::ResizeViewports()
     }
     else
     {
-        ActiveViewportClient->GetViewport()->ResizeViewport(FRect(0.0f, 72.f, EditorWidth , EditorHeight ));
+        ActiveViewportClient->GetViewport()->ResizeViewport(FRect(0.0f, 72.f, static_cast<float>(EditorWidth) , static_cast<float>(EditorHeight)));
     }
 }
